C++11 std::shared_ptr fixes:
Test for C++11 std::shared_ptr was always failing because std::make_shared was
calling a constructor of 1 argument which did not exist for the struct A. Changed
test code snippet to call default no argument constructor of A.

Once C++11 std::shared_ptr support was fixed, contentDispositionField.cpp and
contentTypeField.cpp would not compile because std::shared_ptr cannot be implicitly
cast to bool (i.e. in a return statement). Added explicit cast to bool.