Add SMTPS with AUTH PLAIN without SASL
GNU SASL is a nice library but comes with its own prerequisites and dependencies.
As IMAP and POP3 are able to work without SASL it seems to me logical to add some authentication support to SMTP too. As these days most of the communication is encrypted it is common to use simple mechanism like AUTH PLAIN, so I have added it.