Fix cross-compilation and MSVC build with OpenSSL
When building for multiple platforms I encountered several issues.
Fist of all MSVC (Windows) build with TLS support through OpenSSL
requires additional library "crypt32".
Secondly, detection of pthread library failed when a cross-compiler
was used (target platform is not identical to source/build platform).
The "FIND_PACKAGE(Threads)" works fine but "FIND_LIBRARY(...)"
either fails or finds pthread library for the wrong platform.
It seems to methat presence of "CMAKE_THREAD_LIBS_INIT" which
is set by the "FIND_PACKAGE(Threads)" test is a sufficiently
reliable to use it to set "VMIME_HAVE_PTHREAD" but I must admit
that this is not really my cup of coffee :-)